+++ es5/main.js
@@ -79,7 +79,8 @@ function () {
       var sp = $SP().auth(user1);
       sp.list("My List", "https://web.si.te").get({...});
       sp.list("Other List"; "http://my.sharpoi.nt/other.directory/").update(...);
-        // let's say we want to use our own function to set the FedAuth cookie
+        // if you use the AddIn method, it may not work – in that case, you'll need to use the FedAuth method as described there: https://github.com/Aymkdn/SharepointPlus/wiki/Using-the-FedAuth-Cookie#javascript-code
+      // let's say we want to use our own function to set the FedAuth cookie
       var sp = $SP().auth({method:'cookie', function() {
         // we need to return the content of our cookie
         // e.g. if it's a FedAuth we'll return 'FedAuth=YAHeZNEZdfnZEfzfzeKnfze…';